Moved to Use FileInfo Objects for Input and Output Files
Using FileInfo objects allows the user to provide a relative path to the input and output files but send the full path to the STDFUFiles library, which expects them.

The error checking and method calling was simplified by pulling default values from Hex2Dfu to Program and the use of the ternary operators.
-GED 2020/11/20