Upgrade to .NET 9.0
- Added build script to document correct build parameters for creating the .NET tool

!!!
When installing hex2dfu as a .NET tool, it MUST be installed using the x86 version of dotnet. This enables hex2dfu to run as a x86 application, which is necessary due to it's dependency on STDDUFiles.dll which is a x86 library.

`%ProgramFiles(x86)%\dotnet\dotnet.exe tool install --global hex2dfu`

-GED 2025/04/23